+++ docs/examples/basic_context/async_context.py
@@ -12,18 +12,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import asyncio
 
-from typing import Callable, Mapping, Optional, Sequence, Union
+from opentelemetry import baggage, trace
+from opentelemetry.sdk.trace import TracerProvider
 
-AttributeValue = Union[
-    str,
-    bool,
-    int,
-    float,
-    Sequence[Union[None, str]],
-    Sequence[Union[None, bool]],
-    Sequence[Union[None, int]],
-    Sequence[Union[None, float]],
-]
-Attributes = Optional[Mapping[str, AttributeValue]]
-AttributesFormatter = Callable[[], Attributes]
+trace.set_tracer_provider(TracerProvider())
+tracer = trace.get_tracer(__name__)
+
+loop = asyncio.get_event_loop()
+
+
+async def async_span(span):
+    with tracer.use_span(span):
+        ctx = baggage.set_baggage("foo", "bar")
+    return ctx
+
+
+async def main():
+    span = tracer.start_span(name="span")
+    ctx = await async_span(span)
+    print(baggage.get_all(context=ctx))
+
+
+loop.run_until_complete(main())