+++ docs/examples/basic_context/child_context.py
@@ -12,18 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from opentelemetry import baggage, trace
 
-from typing import Callable, Mapping, Optional, Sequence, Union
+tracer = trace.get_tracer(__name__)
 
-AttributeValue = Union[
-    str,
-    bool,
-    int,
-    float,
-    Sequence[Union[None, str]],
-    Sequence[Union[None, bool]],
-    Sequence[Union[None, int]],
-    Sequence[Union[None, float]],
-]
-Attributes = Optional[Mapping[str, AttributeValue]]
-AttributesFormatter = Callable[[], Attributes]
+global_ctx = baggage.set_baggage("context", "global")
+with tracer.start_as_current_span(name="root span") as root_span:
+    parent_ctx = baggage.set_baggage("context", "parent")
+    with tracer.start_as_current_span(
+        name="child span", context=parent_ctx
+    ) as child_span:
+        child_ctx = baggage.set_baggage("context", "child")
+
+print(baggage.get_baggage("context", global_ctx))
+print(baggage.get_baggage("context", parent_ctx))
+print(baggage.get_baggage("context", child_ctx))